# api-contract/src/main/openapi/justvotes-v1.yaml
openapi: 3.0.3
info:
  title: JustVotes API
  version: 1.0.0
  description: |
    The authoritative HTTP contract for JustVotes. Poll audit entries are intentionally public and detailed so every participant can verify voting activity; they contain the public identity and selected option.

    ## Cookie contract

    The browser client uses same-origin credentials (`credentials: "same-origin"`), so it sends and stores these cookies automatically. `userID` is the normalized identity value, valid for ten years, `Path=/`, `HttpOnly`, `SameSite=Lax`, and `Secure` when the effective external request protocol is HTTPS. JavaScript must not read or mirror it. `JSESSIONID` is a server-managed session cookie: `Path=/`, session lifetime, `HttpOnly`, `SameSite=Lax`, and `Secure` for effective external HTTPS. `XSRF-TOKEN` is a CSRF bootstrap cookie: `Path=/`, session lifetime, readable by JavaScript (`HttpOnly` absent/false), `SameSite=Lax`, and `Secure` for effective external HTTPS; the client reads it only to send the matching `X-XSRF-TOKEN` header. The application trusts forwarded protocol headers only after deployment behind its trusted proxy; direct untrusted forwarding headers must not be accepted.

    ## Session and CSRF flow

    1. `GET /csrf` first. It returns `{"token":"...","headerName":"X-XSRF-TOKEN"}` and sets the readable `XSRF-TOKEN` cookie. Keep that cookie and send the JSON token in the returned header name for every unsafe request below.
...

...
servers:
  - url: /api/v1
...
paths:
...
  /admin/session:
    get:
      tags:
        - Session
      operationId: session
      security:
        - adminSession: [ ]
      responses:
        '204':
          description: Active session
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
        '401':
          $ref: '#/components/responses/Problem'
...
  /admin/polls:
    get:
      tags:
        - Polls
      operationId: adminPolls
      security:
        - adminSession: [ ]
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '200':
          description: Polls
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Poll'
    post:
      tags:
        - Polls
      operationId: createPoll
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      requestBody:
        $ref: '#/components/requestBodies/CreatePoll'
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '201':
          description: Created
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
        '400':
          $ref: '#/components/responses/Problem'
  '/admin/polls/{pollId}/options':
    parameters:
      - $ref: '#/components/parameters/PollId'
    put:
      tags:
        - Polls
      operationId: replacePollOptions
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      requestBody:
        $ref: '#/components/requestBodies/Options'
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '200':
          description: Updated
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  '/admin/polls/{pollId}/archive':
    parameters:
      - $ref: '#/components/parameters/PollId'
    put:
      tags:
        - Polls
      operationId: archivePoll
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '200':
          description: Transitioned
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  '/admin/polls/{pollId}/restore-from-archive':
    parameters:
      - $ref: '#/components/parameters/PollId'
    put:
      tags:
        - Polls
      operationId: restorePollFromArchive
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '200':
          description: Restored from archive
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  '/admin/polls/{pollId}/reopen':
    parameters:
      - $ref: '#/components/parameters/PollId'
    put:
      tags:
        - Polls
      operationId: reopenPoll
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '200':
          description: Reopened
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  '/admin/polls/{pollId}/restore':
    parameters:
      - $ref: '#/components/parameters/PollId'
    put:
      tags:
        - Polls
      operationId: restorePoll
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '200':
          description: Restored
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  '/admin/polls/{pollId}/publication':
    parameters:
      - $ref: '#/components/parameters/PollId'
    put:
      tags:
        - Polls
      operationId: publishPoll
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      requestBody:
        $ref: '#/components/requestBodies/Expiry'
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '200':
          description: Published
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
    delete:
      tags:
        - Polls
      operationId: makePollPrivate
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '200':
          description: Made private
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  '/admin/polls/{pollId}/expiry':
    parameters:
      - $ref: '#/components/parameters/PollId'
    put:
      tags:
        - Polls
      operationId: changePollExpiry
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      requestBody:
        $ref: '#/components/requestBodies/Expiry'
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '200':
          description: Updated
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  '/admin/polls/{pollId}':
    parameters:
      - $ref: '#/components/parameters/PollId'
    delete:
      tags:
        - Polls
      operationId: deletePoll
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '200':
          description: Deleted
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  '/admin/polls/{pollId}/permanent-deletion':
    parameters:
      - $ref: '#/components/parameters/PollId'
    post:
      tags:
        - Polls
      operationId: permanentlyDeletePoll
      security:
        - adminSession: [ ]
          csrf: [ ]
          csrfCookie: [ ]
      requestBody:
        $ref: '#/components/requestBodies/DeleteConfirmation'
      responses:
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '204':
          description: Permanently deleted
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Server-managed session cookie; Path=/, session lifetime, SameSite=Lax, HttpOnly=true, Secure for effective HTTPS. Sent automatically as same-origin credentials.
    voterIdentity:
      type: apiKey
      in: cookie
      name: userID
      description: Normalized identity cookie; ten-year lifetime, Path=/, SameSite=Lax, HttpOnly=true, Secure for effective HTTPS. Browser-managed and never readable by JavaScript.
    csrf:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: Send the token returned by `GET /csrf` in this header together with its `XSRF-TOKEN` cookie.
    csrfCookie:
      type: apiKey
      in: cookie
      name: XSRF-TOKEN
      description: Path=/, session lifetime, SameSite=Lax, HttpOnly=false, Secure for effective HTTPS. JavaScript reads this cookie and sends the matching token in the X-XSRF-TOKEN header.
  parameters:
    PollId:
      name: pollId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/PollId'
...
  requestBodies:
...
    CreatePoll:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreatePoll'
    Options:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Options'
    Expiry:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Expiry'
...
    DeleteConfirmation:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DeleteConfirmation'
  responses:
    Problem:
      description: Problem Details (RFC 9457). `code` is stable and suitable for machine handling; `detail` is diagnostic text.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  headers:
    NoStore:
      description: Response must not be stored or reused by browsers or intermediary caches.
      schema:
        type: string
        enum: [no-store]
  schemas:
    PollId:
      type: string
      pattern: '^p_v1_[A-Za-z0-9_-]+$'
    GroupId:
      type: string
      pattern: '^g_v1_[A-Za-z0-9_-]+$'
...
    CreatePoll:
      type: object
      additionalProperties: false
      required:
        - title
        - templateGroupId
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '.*\S.*'
        templateGroupId:
          $ref: '#/components/schemas/GroupId'
    Options:
      type: object
      additionalProperties: false
      required:
        - optionTexts
      properties:
        optionTexts:
          type: array
          minItems: 2
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 120
            pattern: '.*\S.*'
    Expiry:
      type: object
      additionalProperties: false
      required:
        - endsAt
      properties:
        endsAt:
          type: string
          format: date-time
          description: UTC timestamp with millisecond precision, for example 2026-08-18T12:00:00.000Z.
...
    DeleteConfirmation:
      type: object
      additionalProperties: false
      required:
        - confirmation
      properties:
        confirmation:
          type: string
          enum:
            - DELETE
...
    TemplateGroup:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - description
      properties:
        id:
          $ref: '#/components/schemas/GroupId'
        name:
          type: string
        description:
          type: string
    Poll:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - visibility
        - state
        - templateGroup
        - templateSnapshotOptions
        - options
      properties:
        id:
          $ref: '#/components/schemas/PollId'
        title:
          type: string
        visibility:
          type: string
        state:
          type: string
        endsAt:
          type: string
          format: date-time
          nullable: true
        templateGroup:
          $ref: '#/components/schemas/TemplateGroup'
        templateSnapshotOptions:
          type: array
          items:
            $ref: '#/components/schemas/Option'
        options:
          type: array
          items:
            $ref: '#/components/schemas/Option'
    Option:
      type: object
      additionalProperties: false
      required:
        - number
        - text
      properties:
        number:
          type: integer
          minimum: 1
        text:
          type: string
...
    Problem:
      type: object
      additionalProperties: true
      required:
        - type
        - title
        - status
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string